Split seek.ts into multiple files
